Implemented Optional attribute on Flags
This allows us to tell the serializer not to touch Flags field in the
basetype payloads. This allows us to grab the 4 bytes which can be used
for something else in the childtype deserialization. For example Flags
may send the length of an array in the child type. So we override using
the property to allow us to steal the 4 bytes to read it as a length
prefixed array.