0x60 is now just a container for commands; Commands have their own handlers
So there was an unsupported serialization scenario that I worked to
support on FreecraftCore.Serializer. That is multi-level inheritance
polymorphic deserialization.

Even though that feature now works it cannot be used for 0x60 command
container payloads because the key data for the Type is expected in the
begining of the payload. Inbetween the two Type keys is 4 byte Flags
from the packet header. Because of this we couldn't deserialize the
heriarhcy properly.

So now there is a command handler for commands and they are a seperate
object that is just stuck inside and contained within a payload.